fix: use correct units for settlement SLI. Increase default settlement deadline to 1.5 hours (#58)
This PR fixes a bug where settlement deadline was 15.000 minutes - not
just 15 minutes.

But 15 minutes is not attainable on e.g. mainnet-dawn where we seal
batches each 1hour. The short-term solution is to increase the deadline
on SLI side to 1.5 hours. We are still discussing longer-term. I updated
the SLIs notion
[page](https://www.notion.so/matterlabs/Measuring-availability-SLIs-and-SLOs-261a48363f2380939067dde8af1af4b6?source=copy_link#2cca48363f238069a427dd68a42d5ecf).